move primitives::Account to primitives-traits::Account
