+++ crates/primitives-traits/src/account.rs
@@ -1,30 +1,47 @@
-use alloy_primitives::{B256, U256};
+use alloy_consensus::constants::KECCAK_EMPTY;
+use alloy_genesis::GenesisAccount;
+use alloy_primitives::{keccak256, B256, U256};
+use reth_codecs::{main_codec, Compact};
 
-/// Represents a minimal Account.
-pub trait Account {
-    /// Get `nonce`.
-    fn nonce(&self) -> u64;
-
-    /// Get `balance`.
-    fn balance(&self) -> U256;
-
-    /// Get `bytecode_hash`.
-    fn bytecode_hash(&self) -> Option<B256>;
+/// An Ethereum account.
+#[main_codec]
+#[derive(Clone, Copy, Debug, PartialEq, Eq, Default)]
+pub struct Account {
+    /// Account nonce.
+    pub nonce: u64,
+    /// Account balance.
+    pub balance: U256,
+    /// Hash of the account's bytecode.
+    pub bytecode_hash: Option<B256>,
 }
 
-impl Account for revm_primitives::AccountInfo {
-    fn nonce(&self) -> u64 {
-        self.nonce
+impl Account {
+    /// Whether the account has bytecode.
+    pub const fn has_bytecode(&self) -> bool {
+        self.bytecode_hash.is_some()
     }
 
-    fn balance(&self) -> U256 {
-        self.balance
+    /// After `SpuriousDragon` empty account is defined as account with nonce == 0 && balance == 0
+    /// && bytecode = None (or hash is [`KECCAK_EMPTY`]).
+    pub fn is_empty(&self) -> bool {
+        self.nonce == 0 &&
+            self.balance.is_zero() &&
+            self.bytecode_hash.map_or(true, |hash| hash == KECCAK_EMPTY)
     }
 
-    fn bytecode_hash(&self) -> Option<B256> {
-        if self.is_empty_code_hash() {
-            return None
+    /// Makes an [Account] from [`GenesisAccount`] type
+    pub fn from_genesis_account(value: &GenesisAccount) -> Self {
+        Self {
+            // nonce must exist, so we default to zero when converting a genesis account
+            nonce: value.nonce.unwrap_or_default(),
+            balance: value.balance,
+            bytecode_hash: value.code.as_ref().map(keccak256),
         }
-        Some(self.code_hash)
+    }
+
+    /// Returns an account bytecode's hash.
+    /// In case of no bytecode, returns [`KECCAK_EMPTY`].
+    pub fn get_bytecode_hash(&self) -> B256 {
+        self.bytecode_hash.unwrap_or(KECCAK_EMPTY)
     }
 }
