+++ crates/primitives/src/account.rs
@@ -1,71 +1,14 @@
 use crate::{
-    keccak256,
     revm_primitives::{Bytecode as RevmBytecode, Bytes},
-    GenesisAccount, B256, KECCAK_EMPTY, U256,
 };
 use byteorder::{BigEndian, ReadBytesExt};
 use bytes::Buf;
-use reth_codecs::{main_codec, Compact};
+use reth_codecs::Compact;
 use revm_primitives::JumpTable;
 use serde::{Deserialize, Serialize};
 use std::ops::Deref;
 
-/// An Ethereum account.
-#[main_codec]
-#[derive(Clone, Copy, Debug, PartialEq, Eq, Default)]
-pub struct Account {
-    /// Account nonce.
-    pub nonce: u64,
-    /// Account balance.
-    pub balance: U256,
-    /// Hash of the account's bytecode.
-    pub bytecode_hash: Option<B256>,
-}
-
-impl Account {
-    /// Whether the account has bytecode.
-    pub const fn has_bytecode(&self) -> bool {
-        self.bytecode_hash.is_some()
-    }
-
-    /// After `SpuriousDragon` empty account is defined as account with nonce == 0 && balance == 0
-    /// && bytecode = None (or hash is [`KECCAK_EMPTY`]).
-    pub fn is_empty(&self) -> bool {
-        self.nonce == 0 &&
-            self.balance.is_zero() &&
-            self.bytecode_hash.map_or(true, |hash| hash == KECCAK_EMPTY)
-    }
-
-    /// Makes an [Account] from [`GenesisAccount`] type
-    pub fn from_genesis_account(value: &GenesisAccount) -> Self {
-        Self {
-            // nonce must exist, so we default to zero when converting a genesis account
-            nonce: value.nonce.unwrap_or_default(),
-            balance: value.balance,
-            bytecode_hash: value.code.as_ref().map(keccak256),
-        }
-    }
-
-    /// Returns an account bytecode's hash.
-    /// In case of no bytecode, returns [`KECCAK_EMPTY`].
-    pub fn get_bytecode_hash(&self) -> B256 {
-        self.bytecode_hash.unwrap_or(KECCAK_EMPTY)
-    }
-}
-
-impl reth_primitives_traits::Account for Account {
-    fn nonce(&self) -> u64 {
-        self.nonce
-    }
-
-    fn balance(&self) -> U256 {
-        self.balance
-    }
-
-    fn bytecode_hash(&self) -> Option<B256> {
-        self.bytecode_hash
-    }
-}
+pub use reth_primitives_traits::Account;
 
 /// Bytecode for an account.
 ///
@@ -149,8 +92,8 @@ impl Compact for Bytecode {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::hex_literal::hex;
-    use revm_primitives::LegacyAnalyzedBytecode;
+    use crate::{hex_literal::hex, KECCAK_EMPTY, U256, B256};
+    use revm_primitives::{LegacyAnalyzedBytecode};
 
     #[test]
     fn test_account() {
